refactor(types): move enums and SearchProductsBody interface to bhoonidhiApiClient.types.ts
