+++ src/utils/types/bhoonidhiApiClient.types.ts
@@ -42,3 +42,40 @@ interface CollectionItem {
 export interface AllCollections {
   Results: CollectionItem[];
 }
+
+export enum ProductType {
+  Standard = 'Standard',
+}
+
+export enum QueryType {
+  shape = 'shape',
+}
+
+export enum IsMX {
+  Yes = 'Yes',
+  No = 'No',
+}
+
+export interface SearchProductsBody {
+  userId: string;
+  prod: ProductType | string;
+  selSats: string;
+  offset: string;
+  sdate: string;
+  edate: string;
+  query: string;
+  queryType: QueryType | string;
+  isMX: IsMX | string;
+  loc?: string;
+  lat?: string;
+  lon?: string;
+  radius?: string;
+  tllat?: string;
+  tllon?: string;
+  brlat?: string;
+  brlon?: string;
+  shpCat?: string;
+  shapefilename?: string;
+  filters: string;
+  [key: string]: string | undefined;
+}
